Factor conflict scenario commonalities into PlanningSequenceScenario
