+++ monitoring/uss_qualifier/scenarios/astm/utm/nominal_planning/conflict_equal_priority_not_permitted/conflict_equal_priority_not_permitted.py
@@ -3,12 +3,8 @@ from uas_standards.astm.f3548.v21.constants import (
     Scope,
 )
 
-from monitoring.monitorlib.clients.flight_planning.client import (
-    FlightPlannerClient,
-)
 from monitoring.monitorlib.clients.flight_planning.flight_info import (
     AirspaceUsageState,
-    FlightInfo,
     UasState,
 )
 from monitoring.monitorlib.clients.flight_planning.flight_info_template import (
@@ -18,22 +14,16 @@ from monitoring.monitorlib.clients.flight_planning.planning import (
     FlightPlanStatus,
     PlanningActivityResult,
 )
-from monitoring.uss_qualifier.resources.astm.f3548.v21 import DSSInstanceResource
-from monitoring.uss_qualifier.resources.astm.f3548.v21.dss import DSSInstance
+from monitoring.uss_qualifier.resources.astm.f3548.v21.dss import DSSInstanceResource
 from monitoring.uss_qualifier.resources.flight_planning import FlightIntentsResource
-from monitoring.uss_qualifier.resources.flight_planning.flight_intent import (
-    FlightIntentID,
-)
 from monitoring.uss_qualifier.resources.flight_planning.flight_intent_validation import (
     ExpectedFlightIntent,
-    estimate_scenario_execution_max_extents,
-    validate_flight_intent_templates,
 )
 from monitoring.uss_qualifier.resources.flight_planning.flight_planners import (
     FlightPlannerResource,
 )
-from monitoring.uss_qualifier.scenarios.astm.utm.clear_area_validation import (
-    validate_clear_area,
+from monitoring.uss_qualifier.scenarios.astm.utm.nominal_planning.planning_sequence_scenario import (
+    PlanningSequenceScenario,
 )
 from monitoring.uss_qualifier.scenarios.astm.utm.test_steps import OpIntentValidator
 from monitoring.uss_qualifier.scenarios.flight_planning.prioritization_test_steps import (
@@ -44,19 +34,17 @@ from monitoring.uss_qualifier.scenarios.flight_planning.prioritization_test_step
 )
 from monitoring.uss_qualifier.scenarios.flight_planning.test_steps import (
     activate_flight,
-    cleanup_flights,
     delete_flight,
     plan_flight,
     submit_flight,
 )
 from monitoring.uss_qualifier.scenarios.scenario import (
     ScenarioCannotContinueError,
-    TestScenario,
 )
 from monitoring.uss_qualifier.suites.suite import ExecutionContext
 
 
-class ConflictEqualPriorityNotPermitted(TestScenario):
+class ConflictEqualPriorityNotPermitted(PlanningSequenceScenario):
     flight1_id: str | None = None
     flight1_planned: FlightInfoTemplate
     flight1_activated: FlightInfoTemplate
@@ -70,22 +58,13 @@ class ConflictEqualPriorityNotPermitted(TestScenario):
     flight2_activated: FlightInfoTemplate
     flight2_nonconforming: FlightInfoTemplate
 
-    tested_uss: FlightPlannerClient
-    control_uss: FlightPlannerClient
-    dss: DSSInstance
-    flight_intents_templates: dict[FlightIntentID, FlightInfoTemplate]
-
     def __init__(
         self,
         tested_uss: FlightPlannerResource,
         control_uss: FlightPlannerResource,
         dss: DSSInstanceResource,
-        flight_intents: FlightIntentsResource | None = None,
+        flight_intents: FlightIntentsResource,
     ):
-        super().__init__()
-        self.tested_uss = tested_uss.client
-        self.control_uss = control_uss.client
-
         scopes = {
             Scope.StrategicCoordination: "search for operational intent references to verify outcomes of planning activities and retrieve operational intent details"
         }
@@ -93,8 +72,6 @@ class ConflictEqualPriorityNotPermitted(TestScenario):
             scopes[Scope.ConformanceMonitoringForSituationalAwareness] = (
                 "query for telemetry for off-nominal operational intents"
             )
-
-        self.dss = dss.get_instance(scopes)
 
         expected_flight_intents = [
             ExpectedFlightIntent(
@@ -164,54 +141,16 @@ class ConflictEqualPriorityNotPermitted(TestScenario):
             ),  # Note: this intent expected to produce Nonconforming state, but this is hard to verify without telemetry.  UAS state is not actually off-nominal.
         ]
 
-        self.flight_intents_templates = (
-            flight_intents.get_flight_intents() if flight_intents else {}
-        )
-        try:
-            validate_flight_intent_templates(
-                self.flight_intents_templates, expected_flight_intents
-            )
-        except ValueError as e:
-            raise ValueError(
-                f"`{self.me()}` TestScenario requirements for flight_intents not met: {e}"
-            )
-
-        for efi in expected_flight_intents:
-            setattr(
-                self,
-                efi.intent_id.replace("equal_prio_", ""),
-                self.flight_intents_templates[efi.intent_id],
-            )
-
-    def resolve_flight(self, flight_template: FlightInfoTemplate) -> FlightInfo:
-        return flight_template.resolve(self.time_context.evaluate_now())
-
-    def run(self, context: ExecutionContext):
-        self.begin_test_scenario(context)
-
-        self.record_note(
-            "Tested USS",
-            f"{self.tested_uss.participant_id}",
-        )
-        self.record_note(
-            "Control USS",
-            f"{self.control_uss.participant_id}",
+        super().__init__(
+            flight_intents=flight_intents,
+            expected_flight_intents=expected_flight_intents,
+            tested_uss=tested_uss,
+            control_uss=control_uss,
+            dss=dss,
+            scopes=scopes,
         )
 
-        self.begin_test_case("Prerequisites check")
-        self.begin_test_step("Verify area is clear")
-        estimated_max_extents = estimate_scenario_execution_max_extents(
-            self.time_context, self.flight_intents_templates
-        )
-        validate_clear_area(
-            self,
-            self.dss,
-            [estimated_max_extents],
-            ignore_self=False,
-        )
-        self.end_test_step()
-        self.end_test_case()
-
+    def run_planning_sequence(self, context: ExecutionContext):
         self.begin_test_case("Attempt to plan flight into conflict")
         self._attempt_plan_flight_conflict()
         self.end_test_case()
@@ -233,8 +172,6 @@ class ConflictEqualPriorityNotPermitted(TestScenario):
         self.begin_test_case("Modify activated flight with pre-existing conflict")
         self._modify_activated_flight_preexisting_conflict(flight_1_oi_ref)
         self.end_test_case()
-
-        self.end_test_scenario()
 
     def _attempt_plan_flight_conflict(self):
         self.begin_test_step("Plan Flight 2")
@@ -536,8 +473,3 @@ class ConflictEqualPriorityNotPermitted(TestScenario):
             }:
                 validator.expect_not_shared()
         self.end_test_step()
-
-    def cleanup(self):
-        self.begin_cleanup()
-        cleanup_flights(self, (self.control_uss, self.tested_uss))
-        self.end_cleanup()
